refactor(api): extract chunk_size auto-sizer to shared module (#427)
`_auto_chunk_size` is going to be reused by the upcoming
`evaluate_chunked`, but its raised `UserInputError` hardcodes
`func_name="run_metrics_chunked"` so a sibling caller would surface
the wrong API in the missing-psutil hint. Move the heuristic +
constants to `factrix/_chunk_size.py`, parameterise `func_name` and
`docs_path`, and keep `_run_metrics._auto_chunk_size` as a thin
wrapper so the existing test import stays meaningful.
