+++ factrix/_run_metrics.py
@@ -849,20 +849,6 @@ def _dispatch_iter(
 
 _DEFAULT_BASE_COLS: tuple[str, ...] = ("date", "asset_id", "forward_return")
 
-# Per-factor in-memory amplification factor used by the chunk-size
-# heuristic. Each chunk's peak RSS is dominated by panel materialise +
-# polars query intermediates (``with_columns(rank_exprs)`` adds one
-# ``_rank__<f>`` column per factor); ``4 ×`` an 8-byte-per-row column
-# slice empirically tracks observed M-ic / S2 peak_rss within ±20%
-# across the small / large presets — close enough for a budget knob.
-_AUTO_CHUNK_OVERHEAD_FACTOR = 4
-
-# Divisor applied to ``psutil.virtual_memory().available`` to derive
-# the per-chunk peak budget. Dividing by 4 (i.e. targeting ~25%) leaves
-# slack for OS, BLAS arenas, and the caller's downstream sink without
-# forcing every batch back to chunk_size=1 on tight machines.
-_AUTO_CHUNK_RSS_DIVISOR = 4
-
 
 def _validate_factor_cols(factor_cols: Sequence[str]) -> list[str]:
     """Shared factor_cols validation for run_metrics + run_metrics_chunked."""
@@ -881,32 +867,17 @@ def _validate_factor_cols(factor_cols: Sequence[str]) -> list[str]:
 
 
 def _auto_chunk_size(n_rows: int, n_factors: int) -> int:
-    """Pick chunk size targeting ~25% of available RAM as per-chunk peak.
-
-    ``psutil`` is an optional dependency (``factrix[bench]`` extras);
-    when absent, callers must pass ``chunk_size`` explicitly — the
-    function raises so the user reaches for the right knob rather
-    than silently getting a degenerate default.
+    """Thin wrapper around the shared helper — keeps the run_metrics caller's
+    ``func_name`` / ``docs_path`` in the missing-psutil hint.
     """
-    try:
-        import psutil  # type: ignore[import-untyped]
-    except ImportError as exc:
-        raise UserInputError(
-            func_name="run_metrics_chunked",
-            field="chunk_size",
-            value=None,
-            expected=(
-                "an explicit positive integer when psutil is not installed; "
-                "auto-sizing requires `pip install psutil` (or "
-                "`pip install 'factrix[bench]'`)"
-            ),
-            docs_path="api/run_metrics_chunked#chunk_size",
-        ) from exc
+    from factrix._chunk_size import auto_chunk_size
 
-    available = psutil.virtual_memory().available
-    per_factor_bytes = max(n_rows * 8 * _AUTO_CHUNK_OVERHEAD_FACTOR, 1)
-    budget = max(available // _AUTO_CHUNK_RSS_DIVISOR, per_factor_bytes)
-    return max(1, min(n_factors, budget // per_factor_bytes))
+    return auto_chunk_size(
+        n_rows,
+        n_factors,
+        func_name="run_metrics_chunked",
+        docs_path="api/run_metrics_chunked#chunk_size",
+    )
 
 
 def run_metrics_chunked(
